chatbot-core/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-community
langchain-openai
langchain-chroma
openai
chromadb
tiktoken